Refactor: hook 함수 분리
