+++ src/components/feature/group/GroupMemberItem.tsx
@@ -3,14 +3,10 @@
 import DropdownSmall from "@/components/ui/DropdownSmall";
 import { EllipsisVertical } from "lucide-react";
 import Image from "next/image";
-import { useState } from "react";
 import crownIcon from "@/assets/icon/crown_icon.svg";
 import ControlledAlertBox from "@/components/ui/ControlledAlertBox";
 import { profileImages } from "@/lib/profileImages";
-import {
-  useRemoveGroupMember,
-  useUpdateMemberPermissions,
-} from "@/lib/api/groupApi";
+import { useGroupMemberActions } from "./members/hooks/useGroupMemberLogic";
 
 interface GroupMemberItemProps {
   profileNum: number;
@@ -31,63 +27,17 @@ const GroupMemberItem = ({
   isLeader,
   groupId,
 }: GroupMemberItemProps) => {
-  const [isOpen, setIsOpen] = useState(false);
-  const [isAlertOpen, setIsAlertOpen] = useState(false);
-  const [alertAction, setAlertAction] = useState<"kick" | "transfer" | "take">(
-    "kick"
-  );
-  const updateMemberPermissions = useUpdateMemberPermissions();
-  const removeMember = useRemoveGroupMember();
-
+  const {
+    isOpen,
+    setIsOpen,
+    isAlertOpen,
+    setIsAlertOpen,
+    handleTopClick,
+    handleBottomClick,
+    getAlertContent,
+    handleAlertAction,
+  } = useGroupMemberActions(groupId);
   const profileImage = profileImages[profileNum];
-
-  const handleTopClick = () => {
-    setAlertAction(role === "GROUP_LEADER" ? "take" : "transfer");
-    setIsAlertOpen(true);
-  };
-
-  const handleBottomClick = () => {
-    setAlertAction("kick");
-    setIsAlertOpen(true);
-  };
-
-  const getAlertContent = () => {
-    switch (alertAction) {
-      case "kick":
-        return `${name}님을 그룹에서 내보내시겠습니까?`;
-      case "transfer":
-        return `${name}님에게 방장 권한을 주시겠습니까?`;
-      case "take":
-        return `${name}님의 방장 권한을 뺏으시겠습니까?`;
-    }
-  };
-
-  const handleAlertAction = () => {
-    switch (alertAction) {
-      case "kick":
-        console.log("멤버 내보내기");
-        removeMember.mutate({
-          groupId,
-          userId: memberId,
-        });
-        break;
-      case "transfer":
-        updateMemberPermissions.mutate({
-          groupId,
-          userId: memberId,
-          groupRole: "GROUP_LEADER",
-        });
-        break;
-      case "take":
-        console.log("방장 권한 뺏기");
-        updateMemberPermissions.mutate({
-          groupId,
-          userId: memberId,
-          groupRole: "GROUP_MEMBER",
-        });
-        break;
-    }
-  };
   return (
     <div className="flex items-center justify-between w-full">
       <div className="flex gap-3 items-center">
@@ -115,7 +65,7 @@ const GroupMemberItem = ({
               <DropdownSmall
                 isOpen={isOpen}
                 onClose={() => setIsOpen(false)}
-                onTopClick={handleTopClick}
+                onTopClick={() => handleTopClick(role)}
                 onBottomClick={handleBottomClick}
               >
                 {role === "GROUP_LEADER"
@@ -125,12 +75,12 @@ const GroupMemberItem = ({
             </div>
           )}
           <ControlledAlertBox
-            content={getAlertContent()}
+            content={getAlertContent(name)}
             cancel="취소"
             action="확인"
             isOpen={isAlertOpen}
             onClose={() => setIsAlertOpen(false)}
-            actionHandler={handleAlertAction}
+            actionHandler={() => handleAlertAction(memberId)}
           />
         </div>
       )}
